// src/app/result/enum/answer-id.enum.ts
export enum AnswerID {
  odkladaniePenaziVyskaDo1x = 'odkladaniePenaziVyskaDo1x',
  odkladaniePenaziVyskaDo3x = 'odkladaniePenaziVyskaDo3x',
  odkladaniePenaziVyskaDo6x = 'odkladaniePenaziVyskaDo6x',
  odkladaniePenaziVyskaNad6x = 'odkladaniePenaziVyskaNad6x',

  kolkoMesacneOstavaNaUcteDo10 = 'kolkoMesacneOstavaNaUcteDo10',
  kolkoMesacneOstavaNaUcteDo30 = 'kolkoMesacneOstavaNaUcteDo30',
  kolkoMesacneOstavaNaUcteNad30 = 'kolkoMesacneOstavaNaUcteNad30',

  kolkoMesacneSporiDo5 = 'kolkoMesacneSporiDo5',
  kolkoMesacneSporiNad5 = 'kolkoMesacneSporiNad5',
  kolkoMesacneSporiNic = 'kolkoMesacneSporiNic',

  zivotnePoistenieMa = 'zivotnePoistenieMa',
  zivotnePoisteniePlatiNiektoIny = 'zivotnePoisteniePlatiNiektoIny',
  zivotnePoistenieNema = 'zivotnePoistenieNema',

  zivotnePoistenieKolkoPlatiPod3 = 'zivotnePoistenieKolkoPlatiPod3',
  zivotnePoistenieKolkoPlati3az5 = 'zivotnePoistenieKolkoPlati3az5',
  zivotnePoistenieKolkoPlatiNad5 = 'zivotnePoistenieKolkoPlatiNad5',

  zivotnePoistenieInvesticiaAno = 'zivotnePoistenieInvesticiaAno',
  zivotnePoistenieInvesticiaNie = 'zivotnePoistenieInvesticiaNie',
  zivotnePoistenieInvesticiaNeviem = 'zivotnePoistenieInvesticiaNeviem',

  majetokBytDomPoistenieVobec = 'majetokBytDomPoistenieVobec',
  majetokBytDomPoistenieNeviemAko = 'majetokBytDomPoistenieNeviemAko',
  majetokBytDomPoistenieKomplexne = 'majetokBytDomPoistenieKomplexne',
  majetokBytDomPoistenieNieKomplexne = 'majetokBytDomPoistenieNieKomplexne',

  investicieAno = 'investicieAno',
  investicieNie = 'investicieNie',

  investicieMesacneDo5 = 'investicieMesacneDo5',
  investicieMesacneDo10 = 'investicieMesacneDo10',
  investicieMesacneNad10 = 'investicieMesacneNad10',

  druhyPilierAno = 'druhyPilierAno',
  druhyPilierNie = 'druhyPilierNie',
  druhyPilierNeviem = 'druhyPilierNeviem',

  druhyPilierFondyIndexove = 'druhyPilierFondyIndexove',
  druhyPilierFondyAkciove = 'druhyPilierFondyAkciove',
  druhyPilierFondyDlhopisove = 'druhyPilierFondyDlhopisove',
  druhyPilierFondyZmiesane = 'druhyPilierFondyZmiesane',
  druhyPilierFondyNeviem = 'druhyPilierFondyNeviem',

  tretiPilierAnoZamestnavatelPrispieva = 'tretiPilierAnoZamestnavatelPrispieva',
  tretiPilierAnoZamestnavatelNeprispieva = 'tretiPilierAnoZamestnavatelNeprispieva',
  tretiPilierNemam = 'tretiPilierNemam',
  tretiPilierNeviem = 'tretiPilierNeviem',

  tretiPilierFondyIndexove = 'tretiPilierFondyIndexove',
  tretiPilierFondyAkciove = 'tretiPilierFondyAkciove',
  tretiPilierFondyDlhopisove = 'tretiPilierFondyDlhopisove',
  tretiPilierFondyZmiesane = 'tretiPilierFondyZmiesane',
  tretiPilierFondyNeviem = 'tretiPilierFondyNeviem',

  uverHypotekaPozickaDo20 = 'uverHypotekaPozickaDo20',
  uverHypotekaPozickaDo30 = 'uverHypotekaPozickaDo30',
  uverHypotekaPozickaNad30 = 'uverHypotekaPozickaNad30',

  hypotekaUrokDo1 = 'hypotekaUrokDo1',
  hypotekaUrok1az2 = 'hypotekaUrok1az2',
  hypotekaUrokViacAko2 = 'hypotekaUrokViacAko2',
  hypotekaUrokNeviem = 'hypotekaUrokNeviem',

  uverPozickaUrokDo5 = 'uverPozickaUrokDo5',
  uverPozickaUrok5az10 = 'uverPozickaUrok5az10',
  uverPozickaUrokViacAko10 = 'uverPozickaUrokViacAko10',
  uverPozickaUrokNeviem = 'uverPozickaUrokNeviem',
}